add feature delete for admin

+++ src/main/java/iuh/fit/dhktpm117ctt/group06/service/impl/BrandServiceImpl.java
@@ -11,6 +11,7 @@ import iuh.fit.dhktpm117ctt.group06.repository.BrandRepository;
 import iuh.fit.dhktpm117ctt.group06.service.BrandService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -46,6 +47,7 @@ public class BrandServiceImpl implements BrandService {
     }
 
     @Override
+    @PreAuthorize("hasRole('ADMIN')")
     public Optional<BrandResponse> save(BrandRequest brandRequest) {
         if (brandRequest.getAvatar() == null) {
             return Optional.empty();
@@ -63,6 +65,7 @@ public class BrandServiceImpl implements BrandService {
     }
 
     @Override
+    @PreAuthorize("hasRole('ADMIN')")
     public Optional<BrandResponse> update(String id,BrandRequest brandRequest) {
         Optional<Brand> brand = brandRepository.findById(id);
 
@@ -95,8 +98,11 @@ public class BrandServiceImpl implements BrandService {
 
 
     @Override
-
+//    @PreAuthorize("hasRole('ADMIN')")
     public void deleteById(String id) {
+        if (!brandRepository.checkBrandNotProduct(id)) {
+            throw new AppException(ErrorCode.BRAND_HAS_PRODUCT);
+        }
         brandRepository.deleteById(id);
     }
